+++ src/front/asm/assemble_instruction.cpp
@@ -449,21 +449,25 @@ viua::internals::types::bytecode_size assemble_instruction(Program& program, viu
         TokenIndex source = target + 2;
         TokenIndex position = source + 2;
 
+        int_op target_op, source_op, position_op;
+
         if (tokens.at(target) == "void") {
             --source;
             --position;
-            program.opvpop(
-                assembler::operands::getint(resolveregister(tokens.at(target)))
-                , assembler::operands::getint_with_rs_type(resolveregister(tokens.at(source)), resolve_rs_type(tokens.at(source+1)))
-                , assembler::operands::getint(resolveregister(tokens.at(position), true), true)
-            );
+            target_op = assembler::operands::getint(resolveregister(tokens.at(target)));
         } else {
-            program.opvpop(
-                assembler::operands::getint_with_rs_type(resolveregister(tokens.at(target)), resolve_rs_type(tokens.at(target+1)))
-                , assembler::operands::getint_with_rs_type(resolveregister(tokens.at(source)), resolve_rs_type(tokens.at(source+1)))
-                , assembler::operands::getint(resolveregister(tokens.at(position), true), true)
-            );
+            target_op = assembler::operands::getint_with_rs_type(resolveregister(tokens.at(target)), resolve_rs_type(tokens.at(target+1)));
         }
+
+        source_op = assembler::operands::getint_with_rs_type(resolveregister(tokens.at(source)), resolve_rs_type(tokens.at(source+1)));
+
+        if (tokens.at(position) == "void") {
+            position_op = assembler::operands::getint(resolveregister(tokens.at(position)));
+        } else {
+            position_op = assembler::operands::getint_with_rs_type(resolveregister(tokens.at(position)), resolve_rs_type(tokens.at(position+1)));
+        }
+
+        program.opvpop(target_op, source_op, position_op);
     } else if (tokens.at(i) == "vat") {
         TokenIndex target = i + 1;
         TokenIndex source = target + 2;
@@ -472,7 +476,7 @@ viua::internals::types::bytecode_size assemble_instruction(Program& program, viu
         program.opvat(
             assembler::operands::getint_with_rs_type(resolveregister(tokens.at(target)), resolve_rs_type(tokens.at(target+1)))
             , assembler::operands::getint_with_rs_type(resolveregister(tokens.at(source)), resolve_rs_type(tokens.at(source+1)))
-            , assembler::operands::getint(resolveregister(tokens.at(position), true), true)
+            , assembler::operands::getint_with_rs_type(resolveregister(tokens.at(position)), resolve_rs_type(tokens.at(position+1)))
         );
     } else if (tokens.at(i) == "vlen") {
         TokenIndex target = i + 1;